refactor(snowflake): fold injection audit into standard tests; dedupe deny escaping
Consolidates the standalone test_snowflake_query_injection_audit.py into the
standard test_snowflake_query_optimizations.py, keeping only coverage not already
present (and dropping duplicates now that the standard oracle is Snowflake-accurate):

- migrate TestMakeComposableFastPath: locks the re.compile-skip fast path's
  safety contract (subset admits no SQL/regex breaker; fast==slow; literals
  always well-formed)
- migrate TestDenyPatternInjection: deny-path NOT RLIKE literals stay
  self-contained and deny patterns are never silently dropped
- add the missing trailing-backslash paginator case and brace/backslash schema
  template params
- add one oracle-integrity guard so the Snowflake-accurate parser can't regress
  to ''-only

Also folds the deny path onto the shared _escape_sql_string_literal helper so all
four escape sites share one audited implementation. Behavior-preserving.
